backend/requirements.txt
langchain
langchain-anthropic
langchain-openai
langsmith
fastapi
uvicorn
pinecone
python-dotenv
pandas 
langchain-community
pinecone
openpyxl
sentence-transformers
langchain-community
langsmith
anthropic